requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
importlib-metadata==8.2.0
Markdown==3.6
sqlparse==0.5.1
typing-extensions==4.12.2
zipp==3.20.0
